requirements.txt
fastapi==0.111.0
uvicorn==0.34.3
jinja2==3.1.3
openai==1.84.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.11.5
reportlab==4.1.0
python-docx==1.1.0
email-validator==2.1.0.post1
#smtplib  # (built-in but add as comment)
gunicorn
stripe
sendgrid
python-multipart
aiofiles
pdfkit
itsdangerous==2.1.2